requirements.txt
# Core libraries
llama-index==0.10.68
llama-hub
llama-agents==0.0.14
llama-index-vector-stores-pinecone==0.1.2

# Speech recognition
SpeechRecognition
# pyaudio

# Data processing
numpy
pandas

# API and web framework (if needed)
flask

# Testing
pytest

# Configuration
pyyaml

openai
langchain
langgraph
torch
langchain-openai
python-dotenv
fpdf
pinecone-client
PyPDF2
pinecone

# Simulator interface (placeholder, replace with actual Gazebo interface library)
# gazebo-interface==1.0.0